requirements.txt
app_utils==0.1a4
Flask==2.0.2
gevent==21.12.0
Keras==2.2.3
matplotlib==3.4.2
numpy==1.20.3
Pillow==9.0.1
scikit_image==0.19.1
scipy==1.4.1
skimage==0.0
tensorflow==2.2.0
tensorflow_datasets==4.5.2
tqdm==4.62.3
